Removed hpp files from being added to the project by premake. Renamed Application.hpp to be a .h. Corrected #includes.
