+++ src/main.cpp
@@ -1,4 +1,4 @@
-#include "Application.hpp"
+#include "Application.h"
 
 #include <iostream>
